requirements.txt
chainlit==2.5.5
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-google-genai==2.0.10
langchain-ollama==0.3.3
google-cloud-storage==3.1.1
pandas==2.2.3
duckdb==1.3.1
plotly==6.1.2
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0
asyncpg==0.30.0
pyngrok==7.2.11
python-multipart<0.0.19
nltk==3.9.1
groq==0.29.0
pandas
pydantic
nest_asyncio
python-dotenv
requests
fastapi_sessions
chardet
openpyxl

